refactor program code
