+++ programs/aquachain/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "twopart"
+name = "aquachain"
 version = "0.1.0"
 description = "Created with Anchor"
 edition = "2021"
 
 [lib]
 crate-type = ["cdylib", "lib"]
-name = "twopart"
+name = "aquachain"
 
 [features]
 default = []
