+++ programs/aquachain/src/state/mod.rs
@@ -1,0 +1,7 @@
+mod aquifer;
+mod consumer;
+mod tariff;
+
+pub use aquifer::*;
+pub use consumer::*;
+pub use tariff::*;
